refactor: 운동 업데이트 로직 검증 이동
